Add documentation.
Per request of a friend, this adds more explicit and clear documentation
about what this plugin does and how it does it. While I feel the source
is simple enough to read, there are of course many users who will not
want to do so, this makes the project much more accessible to all.