Add testing suite.
The benefits of automated testing are well discussed, I don't need to
repeat that here. Further, the odds are low that I will be adding more
features to this plugin, it does what I want. However, this goes a long
way towards the educational nature of this plugin, both for me and
hopefully for others.

I've learned a lot writing this plugin, and it is my belief that what it
does is simple, while still demonstrating best practices to be an
educational tool. To this end, there may be further commits updating and
expanding comments. Once I am comfortable with how things lie, I will
declare a version 1.0.